# Start of new cross platform regression tests with Python