--- a/configure.ac
+++ b/configure.ac
@@ -3225,6 +3225,8 @@ AC_CONFIG_FILES(Makefile
   tests/unit_tests/Makefile
   tests/unit_tests/lib/Makefile
   tests/unit_tests/lib/common/Makefile
+  tests/regression_tests/Makefile
+  tests/regression_tests/shapes/Makefile
 	share/Makefile
 	share/examples/Makefile
 	share/gui/Makefile