--- a/tests/regression_tests/Makefile.am
+++ b/tests/regression_tests/Makefile.am
@@ -1,0 +1,1 @@
+SUBDIRS = shapes