--- a/tests/regression_tests/regression_tests.bat
+++ b/tests/regression_tests/regression_tests.bat
@@ -1,0 +1,4 @@
+@ECHO OFF
+
+cd shapes
+python shapes.py