--- a/tests/regression_tests/shapes/Makefile.am
+++ b/tests/regression_tests/shapes/Makefile.am
@@ -1,0 +1,2 @@
+check test rtest:
+	python shapes.py