--- a/tests/regression_tests/shapes/reference/Mdiamond.xdot
+++ b/tests/regression_tests/shapes/reference/Mdiamond.xdot
@@ -1,0 +1,14 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 4 27 36 0 18 27 0 54 18 c 7 -#000000 L 2 9 24 9 12 c 7 -#000000 L 2 18 6 36 6 c 7 -#000000 L 2 45 12 45 24 c 7 -#\
+000000 L 2 36 30 18 30 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=Mdiamond,
+		width=0.75];
+}