--- a/tests/regression_tests/shapes/reference/assembly.xdot
+++ b/tests/regression_tests/shapes/reference/assembly.xdot
@@ -1,0 +1,14 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 4 15 19.5 39 19.5 39 25.5 15 25.5 c 7 -#000000 p 4 15 10.5 39 10.5 39 16.5 15 16.5 c 7 -#000000 L 2 39 18 54 18 c \
+7 -#000000 L 2 0 18 15 18 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=assembly,
+		width=0.75];
+}