--- a/tests/regression_tests/shapes/reference/box3d.xdot
+++ b/tests/regression_tests/shapes/reference/box3d.xdot
@@ -1,0 +1,14 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 6 54 36 4 36 0 32 0 0 50 0 54 4 c 7 -#000000 L 2 50 32 0 32 c 7 -#000000 L 2 50 32 50 0 c 7 -#000000 L 2 50 32 54 \
+36 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=box3d,
+		width=0.75];
+}