--- a/tests/regression_tests/shapes/reference/doublecircle.xdot
+++ b/tests/regression_tests/shapes/reference/doublecircle.xdot
@@ -1,0 +1,13 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 44 44 44 44 0 ",
+		bb="0,0,44,44",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 e 22 22 18 18 c 7 -#000000 e 22 22 22 22 ",
+		height=0.61111,
+		label="",
+		pos="22,22",
+		shape=doublecircle,
+		width=0.61111];
+}