--- a/tests/regression_tests/shapes/reference/insulator.xdot
+++ b/tests/regression_tests/shapes/reference/insulator.xdot
@@ -1,0 +1,14 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 4 33 24 33 12 21 12 21 24 c 7 -#000000 L 5 36 27 36 9 18 9 18 27 36 27 c 7 -#000000 L 2 36 18 54 18 c 7 -#000000 \
+L 2 0 18 18 18 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=insulator,
+		width=0.75];
+}