--- a/tests/regression_tests/shapes/reference/invtriangle.gv
+++ b/tests/regression_tests/shapes/reference/invtriangle.gv
@@ -1,0 +1,9 @@
+graph G {
+	graph [bb="0,0,54,36"];
+	node [label="\N"];
+	a	 [height=0.5,
+		label="",
+		pos="27,18",
+		shape=invtriangle,
+		width=0.75];
+}