--- a/tests/regression_tests/shapes/reference/note.xdot
+++ b/tests/regression_tests/shapes/reference/note.xdot
@@ -1,0 +1,13 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 5 48 36 0 36 0 0 54 0 54 30 c 7 -#000000 L 2 48 36 48 30 c 7 -#000000 L 2 54 30 48 30 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=note,
+		width=0.75];
+}