--- a/tests/regression_tests/shapes/reference/octagon.xdot
+++ b/tests/regression_tests/shapes/reference/octagon.xdot
@@ -1,0 +1,13 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 8 54 10.54 54 25.46 38.18 36 15.82 36 0 25.46 0 10.54 15.82 0 38.18 0 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=octagon,
+		width=0.75];
+}