--- a/tests/regression_tests/shapes/reference/plain.xdot
+++ b/tests/regression_tests/shapes/reference/plain.xdot
@@ -1,0 +1,12 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 0 0 0 0 0 ",
+		bb="0,0,0,0",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [height=0,
+		label="",
+		pos="0,0",
+		shape=plain,
+		width=0];
+}