--- a/tests/regression_tests/shapes/reference/point.gv
+++ b/tests/regression_tests/shapes/reference/point.gv
@@ -1,0 +1,9 @@
+graph G {
+	graph [bb="0,0,3.6,3.6"];
+	node [label="\N"];
+	a	 [height=0.05,
+		label="",
+		pos="1.8,1.8",
+		shape=point,
+		width=0.05];
+}