--- a/tests/regression_tests/shapes/reference/primersite.xdot
+++ b/tests/regression_tests/shapes/reference/primersite.xdot
@@ -1,0 +1,13 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 5 39 21 27 33 27 27 13.5 27 13.5 21 c 7 -#000000 L 2 0 18 54 18 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=primersite,
+		width=0.75];
+}