--- a/tests/regression_tests/shapes/reference/proteinstab.xdot
+++ b/tests/regression_tests/shapes/reference/proteinstab.xdot
@@ -1,0 +1,14 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 8 28.5 24 30 25.5 30 28.5 28.5 30 25.5 30 24 28.5 24 25.5 25.5 24 c 7 -#000000 L 2 27 24 27 18 c 7 -#000000 L 2 0 \
+18 54 18 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=proteinstab,
+		width=0.75];
+}