--- a/tests/regression_tests/shapes/reference/restrictionsite.xdot
+++ b/tests/regression_tests/shapes/reference/restrictionsite.xdot
@@ -1,0 +1,14 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 8 39.75 21 20.25 21 20.25 27 14.25 27 14.25 15 33.75 15 33.75 9 39.75 9 c 7 -#000000 L 2 0 18 14.25 18 c 7 -#000000 \
+L 2 39.75 18 54 18 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=restrictionsite,
+		width=0.75];
+}