--- a/tests/regression_tests/shapes/reference/signature.xdot
+++ b/tests/regression_tests/shapes/reference/signature.xdot
@@ -1,0 +1,13 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 4 54 30 0 30 0 6 54 6 c 7 -#000000 L 2 3 19.5 6 16.5 c 7 -#000000 L 2 3 16.5 6 19.5 c 7 -#000000 L 2 3 9 51 9 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=signature,
+		width=0.75];
+}