--- a/tests/regression_tests/shapes/reference/threepoverhang.xdot
+++ b/tests/regression_tests/shapes/reference/threepoverhang.xdot
@@ -1,0 +1,13 @@
+graph G {
+	graph [_draw_="c 9 -#fffffe00 C 7 -#ffffff P 4 0 0 0 36 54 36 54 0 ",
+		bb="0,0,54,36",
+		xdotversion=1.7
+	];
+	node [label="\N"];
+	a	 [_draw_="c 7 -#000000 p 4 54 19.5 54 25.5 30 25.5 30 19.5 c 7 -#000000 p 4 42 10.5 42 16.5 30 16.5 30 10.5 c 7 -#000000 L 2 0 18 30 18 ",
+		height=0.5,
+		label="",
+		pos="27,18",
+		shape=threepoverhang,
+		width=0.75];
+}