The current regression tests only run on Unix and the reference files aren't up to date. In this pull request I propose a new way to perform the regression tests: using Python. With Python, the tests can be executed on any platform, so the Windows binaries can be tested as well.
Currently the only the different shapes are tested and only with `.gv` output, but I thought it would be a good idea to submit this before spending more time on it.

The reason that I only includes the `.gv` output, is that other output formats produced slightly different results on the different platform. These differences were mostly in the floating point positions of some shapes. It would be best if these differences could be eliminated, where are the exact coordinates of the various shapes determined?