@ellson 2016-10-31
Shape generation is in lib/common/shapes.c

The shapes are sized to fit around labels,   so any variation in fonts on different platforms can have an effect on the layout.    Maybe we should include a font just for regression testing?