@ErwinJanssen 2016-10-31
These regression tests are for shapes without labels, and the differences are very subtle. For example, (part of) the SVG output for the shape "diamond" on Windows:

```
<g id="node1" class="node">
<title>a</title>
<polygon fill="none" stroke="#000000" points="27,-36 0,-18 27,-3.55271e-15 54,-18 27,-36"/>
</g>
```
And on Linux:
```
<g id="node1" class="node">
<title>a</title>
<polygon fill="none" stroke="#000000" points="27,-36 3.55271e-15,-18 27,-3.55271e-15 54,-18 27,-36"/>
</g>
```

All differences in the regression test results are with coordinates that are (close to) zero. Most of them, like the example above, are with one platform setting a coordinate to zero, the other to a value very close to zero (`3.55271e-15`). The rest is one platform setting the coordinate to zero, the other to negative zero.

I think it has something to to with orientation attribute of the shapes, but I couldn't find where the exact coordinates of the shapes are calculated. Adding an edge case for values close to zero should solve it.