@emden 2016-11-02
Testing has always been a bane for us. As far as I can tell, comparing two images has not been solved yet, at least, Google hasn't. Except when the layout fails or there are really egregious differences, you basically have to compare the two images visually. On the other hand, I suspect the solution is not too far away, given the massive amount of work on facial and pattern recognition and the recent advances in AI.

The only hope of some automation was to have separate test data for each machine/OS/version. This usually boiled down to picking one of our machines and only testing on that.

The coordinates for polygonal shapes are generated in lib/common/shapes.c, starting at line 2122, right after the FIXME comment:

        vertices = N_NEW(outp * sides, pointf);
        if (ND_shape(n)->polygon->vertices) {
            poly_desc_t* pd = (poly_desc_t*)ND_shape(n)->polygon->vertices;
            pd->vertex_gen (vertices, &bb);
            xmax = bb.x/2;
            ymax = bb.y/2;
        } else {
            sectorangle = 2. * M_PI / sides;
            sidelength = sin(sectorangle / 2.);
            skewdist = hypot(fabs(distortion) + fabs(skew), 1.);
            gdistortion = distortion * SQRT2 / cos(sectorangle / 2.);
            gskew = skew / 2.;
            angle = (sectorangle - M_PI) / 2.;
            sincos(angle, &sinx, &cosx);
            R.x = .5 * cosx;
            R.y = .5 * sinx;
            xmax = ymax = 0.;
            angle += (M_PI - sectorangle) / 2.;
            for (i = 0; i < sides; i++) {
etc.
Given floating point arithmetic is involved, it is not surprising that subtle differences occur.