@ErwinJanssen 2016-11-03
Visually comparing two images is tricky, but, you created the diffimg utility to solve this problem, right? It's easier with text based output such as gv and SVG, so I started with that first. 

I hope to create some tests that run equally on all platforms, but sometimes this will require changes in the code. I found the location where the coordinates are calculated are added a check for near zero cases. This allowed the regression test to include the SVG and xdot output types (and probably other text based output types). The tests succeed on all platforms I have tested them on: Windows, Debian, Appveyor and Travis.