@magneticnorth 2016-11-03
Yes, agree. Various sources claim to have useful software for computing
“perceptual differences” but it would take some work to figure out if they
are suitable for us.  I don’t know any that deal with possible issues like 
offsets due to different margins or maybe parts of the layout shifted.
http://pdiff.sourceforge.net <http://pdiff.sourceforge.net/>

imagemagick has a powerful image compare utility, and maybe some way
to cough up a metric that reflects the number of pixels that changed.
http://www.imagemagick.org/Usage/compare/ <http://www.imagemagick.org/Usage/compare/>

Sorry not to add more useful information but maybe someone else can help.