@ellson 2016-11-10
Emden, Erwin,

Is this  ready for merging?     If its additional and doesn't affect Emden's current regressions tests, then I'll merge.   Can you please confirm,?