@ErwinJanssen 2016-11-11
These tests are additional and to not influence the current regression tests in any way. They are included in the `make check` step, but if the additional comparisons are added, this step succeeds.

You raise a good point about the additional comparisons performance, I did not check that before sending this PR. To test this I executed `time ./rtest.sh` 15 times with the code before the change and 15 times after. Results in the tables below:

__Before the additional comparisons__

| real time | user time | system time |
| -------- | ---------- | ------- |
| 17.277 | 11.748 | 0.380 |
| 18.284 | 11.868 | 0.648 |
| 17.610 | 11.672 | 0.340 |
| 17.934 | 11.740 | 0.392 |
| 17.774 | 11.736 | 0.404 |
| 17.416 | 11.644 | 0.344 |
| 17.873 | 11.764 | 0.392 |
| 18.161 | 11.788 | 0.380 |
| 17.473 | 11.536 | 0.316 |
| 17.879 | 11.768 | 0.412 |
| 18.112 | 11.716 | 0.456 |
| 17.845 | 11.808 | 0.408 |
| 17.618 | 11.596 | 0.424 |
| 18.114 | 11.872 | 0.348 |
| 17.379 | 11.444 | 0.420 |
| __Averages__  |
| 17.783 | 11.713 | 0.404 |

__After the additional comparisons__

| real time | user time | system time |
| -------- | ---------- | ------- |
| 17.269 | 11.440 | 0.436 |
| 17.813 | 11.844 | 0.408 |
| 17.721 | 11.616 | 0.404 |
| 17.520 | 11.624 | 0.360 |
| 17.574 | 11.540 | 0.432 |
| 17.644 | 11.680 | 0.372 |
| 17.578 | 11.592 | 0.380 |
| 17.612 | 11.612 | 0.408 |
| 17.670 | 11.640 | 0.404 |
| 17.873 | 11.696 | 0.384 |
| 17.238 | 11.428 | 0.400 |
| 17.597 | 11.776 | 0.364 |
| 17.318 | 11.544 | 0.408 |
| 17.405 | 11.588 | 0.400 |
| 17.579 | 11.660 | 0.416 |
| __Averages__  |
| 17.561 | 11.619 | 0.398 |

Looking at this benchmark, I don't see a performance impact.

I'm not sure what you mean with "clamping", could you elaborate?