@emden 2016-11-11
As an aside, I hope to clean up the last of the regressions this weekend, so we should be able to cut a new release next week.